Move BenefitMarkets' Configuration class to module
