--- components/benefit_markets/app/models/benefit_markets/configuration.rb
@@ -1,8 +1,0 @@
-module BenefitMarkets
-  class Configuration
-    include Mongoid::Document
-    include Mongoid::Timestamps
-
-    embedded_in :configurable, polymorphic: true
-  end
-end
