+++ components/benefit_markets/app/models/benefit_markets/configurations/configuration.rb
@@ -1,0 +1,10 @@
+module BenefitMarkets
+	module Configurations
+	  class Configuration
+	    include Mongoid::Document
+	    include Mongoid::Timestamps
+
+	    embedded_in :benefit_market, inverse_of: 'configuration'
+	  end
+	end
+end
