+++ components/benefit_markets/app/views/benefit_markets/benefit_markets/index.html.erb
@@ -12,7 +12,7 @@
     </tr>
   </thead>
   <tbody>
-    <% @benifit_markets.each do |bm| %>
+    <% @benefit_markets.each do |bm| %>
       <tr>
         <td><%= bm.site_urn %></td>
         <td><%= bm.kind %></td>
