+++ components/benefit_markets/app/views/benefit_markets/benefit_markets/new.html.erb
@@ -1,4 +1,4 @@
-<h2>New Benifit Market</h2>
+<h2>New Benefit Market</h2>
 
 <%= form_for @benefit_market, url: { action: 'new'}, html: { data: { controller: 'benefit-market--form' } } do |f| %>
   <%= render partial: 'form', locals: { f: f } %>
